New unsafe abstractions
Add a new unsafe array access helper that eliminates much of the
overhead of array (avoiding requesting bounds to compute an index).
This is safe for the newtyped indices that always start at zero.  Right
now, this is just Variable and Literal (and our temporary use of Int).
Eligible types have to implement an unsafe variant of Ix.

Quick profiling runs indicate it does help non-trivially.

The larger low hanging fruit is the allocation of Seq chunks in the watchlist.