Implement a custom array library
The standard array library is usable, but the range computations are not
really necessary for this use case, given that all of my indices start
from zero.

I wanted an API a bit more like Vector, but supporting index types that
are newtypes of Int (or isomorphic to it).  This array implementation is
built on top of a PrimMonad class, allowing mutable vectors to be used
in either ST or IO.  Furthermore, the abstraction for unboxed data is
easier to use than that in array or vector: it can be derived with
GeneralizedNewtypeDeriving.  That said, it is more efficient to define
Unbox instances manually to include INLINE pragmas.

These arrays are built on the primitives provided by GHC.