Widespread type changes
This essentially replaces almost all instances of IO with the Solver
monad type.  The Solver is now an instance of PrimMonad, so it can use
mutable arrays directly.  This seems to be low to possibly zero
overhead, and it removes a large number of silly liftIO calls.  It also
makes it easier to reuse code and simplifies iteration over mutable arrays.