Implement clause simplification
After backjumping to decision level 0, it is safe to remove clauses that
contain a True literal (they will always be satisfied, and contribute
nothing).  It is also safe to remove False literals from clauses (again,
they will always be false).  If a clause is reduced to a single
unassigned literal, it can be asserted immediately.