Switch to a mutable clause representation
Immutable clauses were very nice conceptually, but the representation
made it difficult to implement new constraint types (e.g., equality,
disequality, or m-of-n constraints).  The new clauses keep all of the
metadata about clauses inline with the clause literals (including
watched literals).  Given this representation, it was simpler to just
make clauses mutable so that the activity and watchlists could be
updated in-place.

This reduces allocation and makes simplification more efficient.